Moved camera and removed dev buttons
