--- app/src/main/java/com/github/bgabriel998/softwaredevproject/Button4Activity.java
@@ -1,19 +1,0 @@
-package com.github.bgabriel998.softwaredevproject;
-
-import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
-
-
-public class Button4Activity extends AppCompatActivity {
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_button4);
-    }
-
-
-
-
-}
