--- app/src/main/java/com/github/bgabriel998/softwaredevproject/Button6Activity.java
@@ -1,14 +1,0 @@
-package com.github.bgabriel998.softwaredevproject;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class Button6Activity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_button6);
-    }
-}
