--- app/src/main/java/com/github/bgabriel998/softwaredevproject/Button8Activity.java
@@ -1,18 +1,0 @@
-package com.github.bgabriel998.softwaredevproject;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-
-public class Button8Activity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_button8);
-
-        Intent intent = new Intent(this, MainMenuActivity.class);
-        startActivity(intent);
-    }
-}
