--- app/src/main/res/layout/activity_button2.xml
@@ -1,70 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/sign_in_main_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/blue_grey_700"
-    android:orientation="vertical"
-    android:weightSum="4"
-    tools:context=".Button2Activity">
-
-    <LinearLayout
-        android:id="@+id/signed_out_layout"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="3"
-        android:gravity="center"
-        android:orientation="vertical">
-
-
-        <TextView
-            android:id="@+id/mountainInfo"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/mountain_info"
-            android:textColor="@android:color/white"
-            android:textSize="14sp" />
-
-        <TextView
-            android:id="@+id/scoreDetail"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:fadeScrollbars="true"
-            android:gravity="center"
-            android:text="Current Score : 0"
-            android:maxLines="5"
-            android:padding="10dp"
-            android:scrollbars="vertical"
-            android:textColor="@android:color/white"
-            android:textSize="14sp" />
-
-
-    </LinearLayout>
-
-
-    <RelativeLayout
-        android:id="@+id/signed_in_layout"
-        android:layout_width="fill_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:background="@color/blue_grey_900">
-
-
-
-        <Button
-            android:id="@+id/compute_user_score_button"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:text="@string/compute_user_score"
-            android:orientation="horizontal"
-            android:paddingLeft="16dp"
-            android:paddingRight="16dp"
-            android:theme="@style/ThemeOverlay.MyDarkButton"
-            tools:visibility="visible" />
-
-
-
-    </RelativeLayout>
-
-</LinearLayout>
