--- app/src/main/res/layout/activity_main.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <Button
-            android:id="@+id/button1"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Camera Preview"
-            android:onClick="button1"/>
-
-        <Button
-            android:id="@+id/button2"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Button2"
-            android:onClick="button2"/>
-
-        <Button
-            android:id="@+id/button3"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Button3"
-            android:onClick="button3"/>
-
-        <Button
-            android:id="@+id/button4"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Get surrounding peaks"
-            android:onClick="button4"/>
-
-        <Button
-            android:id="@+id/button5"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Button5"
-            android:onClick="button5"/>
-
-        <Button
-            android:id="@+id/button6"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Button6"
-            android:onClick="button6"/>
-
-        <Button
-            android:id="@+id/button7"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="Diastance calculator"
-            android:onClick="button7"/>
-
-        <Button
-            android:id="@+id/button8"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:text="App design"
-            android:onClick="button8"/>
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
